requirements.txt
altair==5.5.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.60.1
fqdn==1.5.1
fredapi==0.5.2
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
multitasking==0.0.12
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
streamlit==1.51.0
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
yfinance==0.2.66
